Refactor the converter
Separate the parsing method from the converter by creating a new class and also use the string conversion as the primary method.
